feat(back): ajout des images responsive avec gulp-sharp fix(back): vérification de l'image avant de supprimer dans le dossier local fix(back): fix dockerfile

+++ src/middleware/Images/index.js
@@ -21,6 +21,7 @@ const deletePrevImage = async (req, res, next) => {
   }
   try {
     const [[image]] = await Images.findOneById(req.params.id);
+    if (!image) return res.status(404).send();
     return fs.unlink(path.join(__dirname, `../../../public/images/${image.src}`), (err) => {
       if (err) return res.status(500).send();
       return next();
